學員作品/今天要吃什麼_17蔡忻辰_20陳萱_06林郁雯/main/requirements.txt
ttkthemes
tkinterweb
googlemaps
python-dotenv
pillow